server/chatbot/requirements.txt
streamlit
faiss-cpu
scikit-learn
groq
numpy
